feat(upload): Add posibility to upload multiple collections

+++ src/Traits/Mediable.php
@@ -2,6 +2,8 @@
 
 namespace Shortcodes\Media\Traits;
 
+use Illuminate\Support\Facades\File;
+use Illuminate\Support\Facades\Storage;
 use Spatie\MediaLibrary\HasMedia\HasMediaTrait;
 
 trait Mediable
@@ -25,24 +27,61 @@ trait Mediable
             return $item->name;
         })->toArray();
 
+        $mediaMultipleCollections = collect($this->mediaCollections)->reject(function ($item) {
+            return $item->singleFile === true;
+        })->map(function ($item) {
+            return $item->name;
+        })->toArray();
 
         if (in_array($key, $mediaCollections)) {
 
-            if ($value === null && $this->getFirstMedia($key)) {
-                $this->getFirstMedia($key)->delete();
+            $items = [];
+
+            if (!in_array($key, $mediaMultipleCollections)) {
+                $items[] = ['url' => $value];
             }
 
-            if (strpos($value, '/tmp/') !== false) {
-                $filePath = ltrim(strstr($value, '/tmp/'), '/');
-                $storageFile = storage_path('app/' . $filePath);
+            if (!in_array($key, $mediaMultipleCollections) && $value === null && $this->getFirstMedia($key)) {
+                $this->getFirstMedia($key)->delete();
+                return;
+            }
 
-                $this->addMedia($storageFile)->toMediaCollection($key);
+            if (!$items) {
+                $items = $value;
+            }
+
+            foreach ($items as $item) {
+
+                if (strpos($item['url'], '/tmp/') !== false) {
+                    $filePath = ltrim(strstr($item['url'], '/tmp/'), '/');
+                    $storageFile = storage_path('app/' . $filePath);
+
+                    $media = $this->addMedia($storageFile);
+                    if ($this->isVideo($storageFile)) {
+                        $media->withCustomProperties(['isVideo' => true]);
+                    }
+
+                    $media->toMediaCollection($key);
+                }
             }
 
             return;
         }
 
         return parent::setAttribute($key, $value);
+    }
+
+    private function isVideo($filePath)
+    {
+        $videoMimes = config('upload.video_mimetypes');
+
+        if (!$videoMimes) {
+            return false;
+        }
+
+        $type = File::mimeType($filePath);
+
+        return in_array($type, $videoMimes);
     }
 }
